LPS-124493 Move toForm code to FormUtil and update imports
