+++ modules/apps/headless/headless-form/headless-form-impl/src/main/java/com/liferay/headless/form/internal/resource/v1_0/FormResourceImpl.java
@@ -23,20 +23,17 @@ import com.liferay.dynamic.data.mapping.service.DDMFormInstanceService;
 import com.liferay.headless.form.dto.v1_0.Form;
 import com.liferay.headless.form.dto.v1_0.FormContext;
 import com.liferay.headless.form.dto.v1_0.FormDocument;
-import com.liferay.headless.form.internal.dto.v1_0.util.CreatorUtil;
+import com.liferay.headless.form.dto.v1_0.util.FormDocumentUtil;
+import com.liferay.headless.form.dto.v1_0.util.FormUtil;
 import com.liferay.headless.form.internal.dto.v1_0.util.FormContextUtil;
-import com.liferay.headless.form.internal.dto.v1_0.util.FormDocumentUtil;
-import com.liferay.headless.form.internal.dto.v1_0.util.StructureUtil;
 import com.liferay.headless.form.resource.v1_0.FormResource;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.UserLocalService;
-import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.vulcan.multipart.BinaryFile;
 import com.liferay.portal.vulcan.multipart.MultipartBody;
 import com.liferay.portal.vulcan.pagination.Page;
 import com.liferay.portal.vulcan.pagination.Pagination;
-import com.liferay.portal.vulcan.util.LocalizedMapUtil;
 
 import java.util.Optional;
 
@@ -58,7 +55,10 @@ public class FormResourceImpl extends BaseFormResourceImpl {
 
 	@Override
 	public Form getForm(Long formId) throws Exception {
-		return _toForm(_ddmFormInstanceService.getFormInstance(formId));
+		return FormUtil.toForm(
+			contextAcceptLanguage.isAcceptAllLanguages(),
+			_ddmFormInstanceService.getFormInstance(formId), _portal,
+			contextAcceptLanguage.getPreferredLocale(), _userLocalService);
 	}
 
 	@Override
@@ -70,7 +70,11 @@ public class FormResourceImpl extends BaseFormResourceImpl {
 				_ddmFormInstanceService.getFormInstances(
 					contextCompany.getCompanyId(), siteId,
 					pagination.getStartPosition(), pagination.getEndPosition()),
-				this::_toForm),
+				ddmFormInstance -> FormUtil.toForm(
+					contextAcceptLanguage.isAcceptAllLanguages(),
+					ddmFormInstance, _portal,
+					contextAcceptLanguage.getPreferredLocale(),
+					_userLocalService)),
 			pagination,
 			_ddmFormInstanceService.getFormInstancesCount(
 				contextCompany.getCompanyId(), siteId));
@@ -124,42 +128,6 @@ public class FormResourceImpl extends BaseFormResourceImpl {
 				formDocument.getTitle(), formDocument.getDescription(), null,
 				binaryFile.getInputStream(), binaryFile.getSize(),
 				new ServiceContext()));
-	}
-
-	private Form _toForm(DDMFormInstance ddmFormInstance) throws Exception {
-		if (ddmFormInstance == null) {
-			return null;
-		}
-
-		return new Form() {
-			{
-				availableLanguages = LocaleUtil.toW3cLanguageIds(
-					ddmFormInstance.getAvailableLanguageIds());
-				creator = CreatorUtil.toCreator(
-					_portal,
-					_userLocalService.fetchUser(ddmFormInstance.getUserId()));
-				dateCreated = ddmFormInstance.getCreateDate();
-				dateModified = ddmFormInstance.getModifiedDate();
-				datePublished = ddmFormInstance.getLastPublishDate();
-				defaultLanguage = ddmFormInstance.getDefaultLanguageId();
-				description = ddmFormInstance.getDescription(
-					contextAcceptLanguage.getPreferredLocale());
-				description_i18n = LocalizedMapUtil.getI18nMap(
-					contextAcceptLanguage.isAcceptAllLanguages(),
-					ddmFormInstance.getDescriptionMap());
-				id = ddmFormInstance.getFormInstanceId();
-				name = ddmFormInstance.getName(
-					contextAcceptLanguage.getPreferredLocale());
-				name_i18n = LocalizedMapUtil.getI18nMap(
-					contextAcceptLanguage.isAcceptAllLanguages(),
-					ddmFormInstance.getNameMap());
-				structure = StructureUtil.toFormStructure(
-					contextAcceptLanguage.isAcceptAllLanguages(),
-					ddmFormInstance.getStructure(),
-					contextAcceptLanguage.getPreferredLocale(), _portal,
-					_userLocalService);
-			}
-		};
 	}
 
 	@Reference
